impl rate limiting

+++ src/stream/src/from_proto/chain.rs
@@ -175,7 +175,7 @@ impl ExecutorBuilder for ChainExecutorBuilder {
             ChainType::ChainUnspecified => unreachable!(),
         };
         if let Ok(rate_limit) = node.get_rate_limit() {
-            Ok(FlowControlExecutor::new(executor, *rate_limit as usize).boxed())
+            Ok(FlowControlExecutor::new(executor, *rate_limit).boxed())
         } else {
             Ok(executor)
         }
